- The player now extends the Ball class, inheriting its functionality. - Added functionality to the viewable screen class within the world.

+++ NeonRush/app/src/main/java/alpha/nosleep/neonrush/Ball.java
@@ -35,7 +35,7 @@ public class Ball extends Object implements IPhysics {
         if (img == null)
         {
             Graphics g = getGame().getGraphics();
-            g.drawCircle(position.x, position.y, radius, Color.BLUE);
+            g.drawCircle((int)position.x, (int)position.y, radius, Color.BLUE);
         }
         else
         {
